+++ CHANGELOG.md
@@ -18,3 +18,4 @@ Please, use the format:
 - beta-builder: automatically publish beta packages into pypi from main branch
 - improve debug logging
 - fix issues encountered with PR #11
+- adds a new luxos.cli package